Update ecosystem build tooling  (#901)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** Update Pybind11 to the latest release (2.13.5) for enabling
Numpy 2.0 support, amongst other fixes. In addition, an upgrade to CIBW
is required, as well as a new minimum version for MacOS support to 12.0

**Description of the Change:** Pybind11 updated to v2.13.5. CiBuildWheel
updated to 2.20. MacOS wheel builds updated to a minimum support of
12.0. Allow import failure to return as warning for debugging.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>