Add native PauliRot implementation in LightningKokkos [sc-71642] (#855)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Pauli rotations come up in many places, and importantly in the time
evolution of qchem Hamiltonians. It is therefore worth considering ways
to accelerate their execution.

**Description of the Change:**
Implement `applyPauliRot`. Invoke `applyPauliRot` directly from the SV
class and add bindings to the Python layer.

**Benefits:**
Faster Pauli rotations. I performed a benchmark on random
`PauliRotation`s (runtime > 1.0 sec and at least 5 of them) through the
Python layer. The data remains noisy with 5 samples because the
performance varies depending on the specific "XYZ" sequence (which
translates into more or less predictable memory access patterns).
Overall, we see an advantage for 3+ qubits and up.


![speedup_vs_ntargets_lk_omp16](https://github.com/user-attachments/assets/0fe2bc86-dae2-48c2-9af1-647efe753bae)

I performed the same benchmark on an A100 card with the Kokkos-CUDA
backend, but using at least 500 samples since the absolute timings quite
small and get the following speed-ups.


![speedup_vs_ntargets_lk_cuda](https://github.com/user-attachments/assets/ec842b49-cd93-478b-946e-6f208f97d4de)

Using a full workflow such as 
```
    @qml.qnode(dev, diff_method=None)
    def circuit():
        qml.TrotterProduct(ham, time=1.0, n=1, order=2)
        return qml.state()
```
to benchmark, we obtain timings as follows


![time_vs_mol](https://github.com/user-attachments/assets/7ed7a3db-e71d-42b3-84bb-e0325dceea68)

For large enough molecules (>= 20 qubits, >= 1000 terms), the new
PauliRot kernels have a clear advantage which only grows with molecular
size. It is worth noting that with L-Kokkos-CUDA, even at the (24/10k)
scale, evaluating the circuit is not the main bottleneck which is why it
takes about the same time simulating HCN (2.64 sec. `apply_lightning` vs
32.5 sec. `QNode`) and N2N2 (7.51 sec. `apply_lightning` vs 36.4 sec.
`QNode`).

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-69801]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Luis Alfredo Nuñez Meneses <alfredo.nunez@xanadu.ai>