Prevent labels from triggering CI (#899)
**Context:**
CI is being triggered by adding any label to a PR in "ready for review"
state. This fixes that undesired behaviour

**Description of the Change:**
Removes the `labeled` PR type from workflows. 

**Benefits:**
Prevents unnecessary resource usage.

**Possible Drawbacks:**
When labels are added that intend to affect CI, a followup commit must
be made to trigger the CI

**Related GitHub Issues:**

### Verification:
Added various labels to this PR in a "ready for review" state, no CI
checks were triggered as a result.

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>