Fix qml.probs for lightning.tensor (#906)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

[SC-73501]

`cutensornet`'s `Accessor` API allows the users to extract single state
amplitudes (elements of the state tensor), slices of state amplitudes
(slices of the state tensor) as well as the full state tensor. Previous
implementation summed up the slices of the state tensor and compute the
probabilities, which was a bug. This PR sums up the prob of each slices
of the state tensor and fixes the bug.

**Description of the Change:**

**Benefits:**

Analytic probs is supported

**Possible Drawbacks:**

The computation cost could be too expensive to afford if the number of
projected wires is large.

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>