Downstream LightningSimulator C++ API to the pennylane-lightning repository (#960)
**Context:** Catalyst needs to build a class wrapping the Lightning
Qubit class. Here we are moving the logic from Catalyst to the Lightning
repository.

**Description of the Change:** Moving code from Catalyst to this
repository and updating the build system to build against Catalyst.

**Benefits:** Catalyst wheels will build faster.

**Possible Drawbacks:** Our build system now relies on headers coming
from Catalyst. It is unclear if this may cause a deadlock in the future
because of our cyclic dependencies (PennyLane, PennyLane Lightning,
Catalyst). Chances are small because we don't rely on Catalyst wheels or
build it from scratch.

[sc-70316]
[sc-70318]
[sc-70317]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>
Co-authored-by: paul0403 <79805239+paul0403@users.noreply.github.com>
Co-authored-by: Haochen Wang <paulhaochen.wang@mail.utoronto.ca>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>