Fix the issue with adjoint-jacobian of adjoint ops (#996)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
The Adjoint-Jacobian of adjoint operations in Lightning is incorrectly
returning `0` if we don't decompose them. This issue stems from the ops
serialization logic in Python, which mistakenly assumes `inverse =
False` for all operations. This PR fixes the issue so that we can add
`Adjoint(op)` for all supported `op` to the frozenset of devices
operations that will also fix the blocker for merging PR #988.

``` python

import pennylane as qml

dev = qml.device("lightning.qubit", wires=2)

@qml.qnode(dev, diff_method="adjoint")
def circuit(params):
    qml.adjoint(qml.RX(params[0], wires=[0]))
    qml.adjoint(qml.RY(params[0], wires=[0]))
    qml.adjoint(qml.RZ(params[0], wires=[0]))
    qml.adjoint(qml.SingleExcitation(params[0], wires=[0, 1]))
    return qml.expval(qml.Z(0))

params = qml.numpy.array([0.123])
qml.jacobian(circuit)(params)

```

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-78336]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Joseph Lee <joseph.lee@xanadu.ai>
Co-authored-by: Joseph Lee <40768758+josephleekl@users.noreply.github.com>