(LK-C-4) Add controlled QubitUnitary support to Lightning Kokkos  (#955)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

**Description of the Change:**
This PR adds support for controlled QubitUnitary for Lightning Kokkos.
There are specialized implementations for 1/2/3-qubits
(`applyNC1/2/3QubitOpFunctor`) and a general case implementation
(`NCMultiQubitOpFunctor`). These functors are defined in
`MatrixGateFunctors.hpp`. These are called by `applyControlledMatrix` or
`applyOperation`/`applyNCMultiQubitOp` in `StateVectorKokkos.hpp`

**Benefits:**
Performance benchmarks for gates are shown here:
https://www.notion.so/xanaduai/Lightning-Kokkos-Native-Controlled-Operation-Gate-Benchmarks-12ebc6bd17648017a2dcd237748b24fe

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-76776]