Update Lightning device TOML files to the new schema (#988)
**Context:**

This PR is part of the new device capabilities initiative to improve
feature parity across the ecosystem. See [this
ADR](https://github.com/PennyLaneAI/adrs/pull/78) for more context. A
new TOML schema has been defined, and the relevant module implemented in
PennyLane:
- https://github.com/PennyLaneAI/pennylane/pull/6407
- https://github.com/PennyLaneAI/pennylane/pull/6433.

As well as updates made to Catalyst:
- https://github.com/PennyLaneAI/catalyst/pull/1275

**Description of the Change:**

- Updates `lightning_qubit.toml`, `lightning_kokkos.toml`,
`lightning_gpu.toml` to the new schema
- Removes `_operations` and `_observables` from `LightningQubit`,
`LightningKokkos`, and `LightningGPU` as they are now available via
`Device.capabilities` that is loaded from the TOML file.

**Benefits:**

A step towards feature parity across the ecosystem.

**Possible Drawbacks:**

- Per discussions when the ADR was developed, `operator.gates.decomp`
and `operator.gates.matrix` are removed, and the TOML file no longer
prescribes to the framework how an operator should be handled. To ensure
consistency of behaviour, this information is temporarily moved to a
`_to_matrix_ops` class property to be used by Catalyst, until we have
better support for customizable multi-pathway decompositions.

**Related GitHub Issues:**

[sc-71729]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: Joseph Lee <joseph.lee@xanadu.ai>
Co-authored-by: Joseph Lee <40768758+josephleekl@users.noreply.github.com>