Generalize seeding for measurements (#1003)
**Context:** Currently, only `catalyst.measure()`, `qml.sample()` and
`qml.counts()` can be seeded with qjit(seed=N). It would be nice to be
able to seed all circuit readout methods, like `probs()`, `expval()`,
`var()`, when executing a device with shots.

**Description of the Change:** Let the base measurement class store the
device seed so it can be propagated when necessary to newly created
derived measurements. Move the logic from the simulator to the base
measurement class.

**Benefits:** All measurements using the `generate_samples` method will
be seeded.

[sc-77124]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>