Remove parsing catalyst device attribute dictionary for shots (#1017)
**Context:**
As part of the work to support dynamic measurement shapes,
https://github.com/PennyLaneAI/catalyst/pull/1310 in Catalyst is
allowing a quantum device to take in a dynamic number of shots at the
mlir level. One of the changes involved is that the `DeviceInitOp` now
takes in a proper SSA argument for shots, instead of having shots as
just another entry in the `DeviceInitOp`'s attribute dictionary.

Correspondingly, the backend devices' catalyst interfaces need to stop
parsing the `DeviceInitOp`'s attribute dictionary for shots.

**Description of the Change:**
The backend devices' catalyst interfaces stop parsing the
`DeviceInitOp`'s attribute dictionary for shots.

**Benefits:**
Agreement with Catalyst on how device shots, potentially dynamic, is
handled.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>