Fix hipcc + lightning_kokkos issue (#1018)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [X] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [X] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
This PR fixes an issue with the hipcc (llvm) compiler, which does not
allow `[[maybe_unused]]` attributes in kokkos lambdas, and fails to
correctly resolve `pennylane::utils::exp2` function. E.g. in nightly
rocm docker build
https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/12173013064

**Description of the Change:**
In lightning kokkos, `exp2` are explicitly replaced by
`pennylane::utils::exp2`, and `[[maybe_unused]]` attributes are now
moved from the lambda arguments to the lambda body.

This has been tested to work with the rocm docker build.
(https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/12188525751/job/34001703721)

**Benefits:**
Fixes rocm build failures
**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-79824]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>