Upgrade LK/LGPU CI tests to use cuda 12.4 (#1023)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [X] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Lightning Kokkos CUDA C++ tests fail to compile due to
[bug](https://forums.developer.nvidia.com/t/compiling-a-catch2-application-with-nvcc-std-c-20-leads-a-crash-in-cudafe/255088)
with CUDA 12.2 + Catch2. CI instance is now upgraded to CUDA 12.4, and
all GPU CI tests ([LK, LGPU/LTensor]x[C++, Python]) need to be updated
to load CUDA 12.4.

**Description of the Change:**
Update github workflow to load CUDA 12.4. Also change Lightning Kokkos
C++ test to use `Release` build instead of `RelWithDebInfo` due to
segmentation fault when combined with coverage collection. The relevant
test and line is
[here](https://github.com/PennyLaneAI/pennylane-lightning/blob/e9a3b80823a2dc52f747ff20f91031512d0dbd63/pennylane_lightning/core/src/simulators/lightning_kokkos/gates/tests/Test_StateVectorKokkos_Param.cpp#L2062).

**Benefits:**
Will be able to run Lightning Kokkos C++ tests with CUDA.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-80288]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>