Add the RTLD_NODELETE flag to dlopen (#1030)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

There are segfault issue for finite-shot hermitians tests with Catalyst
on arm64-macos. It looks like closing `libscipy_openblas.dylib` has a
side effect that leads to the segfault and using `RTLD_NODELETE`
mitigates that.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-80901]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>