fix lightning interpreter with for, while, and cond (#1051)
**Context:**

In [PL PR #6786](https://github.com/PennyLaneAI/pennylane/pull/6786), I
added additional kwargs for the for, while, and cond primitives for
handling inputs with abstract shapes. Those inputs broken the custom
registrations for the lightning interpreter.

**Description of the Change:**

Import the `FlattenedHigherOrderPrimitives` to get default behavior for
for, while, and cond. These should better stay updated with changes to
the primitives and the base class.

**Benefits:**

Lightning works again.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>