Optimize MPO operation in LT (#1054)
**Context:**
Reduce the number of `SWAP` gates before applying an MPO to the MPS.  

**Description of the Change:**
The library cuQauntum has support for applying an MPO with unsorted and
non-contiguous target wires.

Internally the library uses the same algorithm as us to perform this
operation to the MPS, i.e. multiple `SWAP` gates to order and make
contiguous the target wires. However, the performance of this internal
algorithm is slightly faster.

**Benefits:**
Simplify our code using the cuQuantum capabilities and increase a bit
the performance of LT for controlled multi-qubit gates.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-82794]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>