Update GPU workflow to use the new self-hosted GPU runner infrastructure (#1061)
**Context:**

Swap the `runs-on` for GPU related workflows to update the infra it uses
to the new GPU infra.

**Description of the Change:**

The current GPU tests use runner infrastructure that is quite old and
have accumulated a fair amount of technical debt. As part of an overhaul
of the infrastructure, net-new instances have been deployed.

This pull request updates the gpu workflows to use the newly deployed
runners.

Though *most* workflows don't require updates, the multi-gpu workflows
do require an update. This is due to the fact that all build
environments have been aligned, and only CUDA 12.4 is available on all
instances.

Due to this the mpi lib, both openmpi and mpich have only been built
against cuda 12.4. And we do not need to load them using
`<mpilib>/cuda-<version>`.

**Benefits:**

Enable us to use newer infrastructure that have non-deprecated runtimes
(backend).

Alignment of multi-gpu build environment to the single gpu build
environment.

**Possible Drawbacks:**

It's possible for a test to fail that used to pass on previous gpu
runners. I have taken steps to ensure the build environment between the
outgoing runners and incoming runners have as much parity as possible.

Of course, we will ensure the gpu tests pass before merging this PR, but
maintainers are encouraged to keep an eye on failing gpu tests over the
next couple days to ensure no failures are happening that are
specifically related to build environment changes.

**Related GitHub Issues:**

[sc-83176](https://app.shortcut.com/xanaduai/story/83176/investigate-and-try-updating-lambda-runtime-for-pl-runners-pennylane-lightning)

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>