[dyn-alloc-base] Lightning dynamically allocate qubit memory (#1043)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
This PR enables lightning devices to dynamically allocate memory for
statevectors just before execution. The usual `qml.device("lightning.x",
wires=y)` is still supported, but user can also not provide `wires`,
i.e. `qml.device("lightning.x")`, and the qubits required for the
statevector is provided by the circuit

**Description of the Change:**
For LQ/LK/LG - the state-vector is only created at the execute step, and
not during device creation. For all devices, if wires is provided during
device creation, this will be the number of wires allocated in the
device state-vector/tensornet; otherwise, it will be determined from the
circuit to be executed.

**Benefits:**

```
dev = qml.device("lightning.qubit")
@qml.qnode(dev)
def circuit():
    qml.Identity(wires=3)
    qml.Identity(wires=1)
    qml.Hadamard(wires=0)
    return qml.state()


@qml.qnode(dev)
def circuit1():
    qml.Identity(wires=1)
    qml.Hadamard(wires=0)
    return qml.state()

circuit()
> array([0.70710678+0.j, 0.70710678+0.j, 0.        +0.j, 0.        +0.j,
       0.        +0.j, 0.        +0.j, 0.        +0.j, 0.        +0.j])
 
circuit1()
> array([0.70710678+0.j, 0.        +0.j, 0.70710678+0.j, 0.        +0.j])
```

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-82702]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>