Fix syntax warning from `is` with a literal (#1070)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
[This PR ](https://github.com/PennyLaneAI/pennylane-lightning/pull/1044)
added two instances of `if tape_return_type is "state"` in
`_adjoint_jacobian_base.py` , that introduces python SyntaxWarning, e.g.
:
```
_adjoint_jacobian_base.py:197: SyntaxWarning: "is" with a literal. Did you mean "=="?
  if tape_return_type is "state":
```
which is also observed in CI logs during testing (e.g.
[here](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/13533320431/job/37820142204#step:14:23)
)

**Description of the Change:**
This PR changes `is` to `==`.

**Benefits:**
No more syntax warning - equivalent test
[here](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/13552303713/job/37878601833?pr=1070#step:14:21)

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-85337]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>