Adding new workflow to allow triggering benchmark from pr (#1073)
**Context:**
Allow us to trigger benchmark from a pull request 

**Description of the Change:**

Adds a net new workflow which would enable us to benchmark a pull
request.

Expected workflow:
- User with at least *write* permissions would need to comment
`/benchmark` on a pull request
- Workflow triggers, checks the comment and users permission using
GitHub API
- If checks pass, a webhook is sent to pennylane-benchmark repo
- The workflow will react to the original comment, indicating that
webhook was sent and the workflow (scoped to lightning) has completed
successfully.

On the Benchmarking side, currently this will trigger a place holder
workflow, which sleeps for a few seconds and replies back with a comment
of "results".

**Benefits:**
Ability to trigger benchmark from a PR.

**Possible Drawbacks:**
This is more scaffolding. The PL benchmarking team would need to update
the workflow that receives this webhook and ensure benchmarks actually
run.

**Related GitHub Issues:**
None.
[sc-85092](https://app.shortcut.com/xanaduai/story/85092/setup-pl-benchmarks-trigger-for-prs-on-pennylane-lightning)

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>