Increase aarch64-cuda wheel build CI max time (#1072)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Due to added python tests in previous, the CI action for aarch64-cuda LT
tests + wheel build as exceeded the maximum time (45min), and are
cancelled resulting in CI error
(https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/13554922529/job/37931172662)

**Description of the Change:**
Increase the max time from 45 to 60 min

**Benefits:**
CI tests and wheel build survives.

**Possible Drawbacks:**


**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>